Only delete .chk files if the option is switched on, even for reactions with code -999

+++ kinbot/reaction_generator.py
@@ -444,9 +444,10 @@ class ReactionGenerator:
                             os.remove('{}_im_extent.txt'.format(self.species.chemid))
                         postprocess.createPESViewerInput(self.species, self.qc, self.par)
                 elif self.species.reac_ts_done[index] == -999:
-                    if not self.species.reac_obj[index].instance_name in deleted:
-                        self.delete_files(self.species.reac_obj[index].instance_name)
-                        deleted.append(self.species.reac_obj[index].instance_name)
+                    if self.par.par['delete_intermediate_files'] == 1:
+                        if not self.species.reac_obj[index].instance_name in deleted:
+                            self.delete_files(self.species.reac_obj[index].instance_name)
+                            deleted.append(self.species.reac_obj[index].instance_name)
                         
             alldone = 1
             for index, instance in enumerate(self.species.reac_inst):
